Create ANOVA_Nested.R
Only datasets::CO2 is usable for nested ANOVAs, but the data is non-normal and has un-equal variances. I'm acting like both assumptions were met to show the steps for this type of analysis. If the viewer can find or generate a dataset that meets both assumptions, they will have the code and explanation to run this analysis.